Move flaky frontend tests to dedicated file (#2604)
